requirements.txt
ursina
boto3
google-generativeai
pyttsx3
sqlalchemy
